requirements/base.txt
django<4.0
django-authtools==2.0.0
django-braces==1.15.0
django-crispy-forms==2.0
crispy-bootstrap3==2022.1
django-environ==0.11.2
pillow==10.1.0
easy-thumbnails==2.8.5
django_select2==8.1.2
psycopg2==2.9.7
django-leaflet==0.28.3
geopy==2.4.0
django-countries==7.5.1
django_summernote==0.8.20.0
django-cookie-law==2.2.0
djangorestframework==3.14.0
djoser==2.2.0
django-rest-swagger==2.2.0
django-oidc-provider==0.8.0
drf-yasg==1.21.7
django-active-link==0.1.8
django-cron==0.5.1
django-ckeditor<6.7
django-oauth-toolkit==2.3.0
whoosh==2.7.4
argon2_cffi==23.1.0
python-dateutil==2.8.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
djangorestframework-csv==2.1.1
django-cleanup==8.0.0
django-admin-tools==0.9.3
cairosvg==2.7.1
Unidecode==1.3.7
requests
beautifulsoup4
